+++ actions/scripts/publish.sh
@@ -56,4 +56,5 @@ else
   deployProject "com.yahoo.athenz:athenz-server-aws-common"
   deployProject "com.yahoo.athenz:athenz-syncer-common"
   deployProject "com.yahoo.athenz:athenz-instance-provider"
+  deployProject "com.yahoo.athenz:athenz-server-msg-pulsar"
 fi
