+++ libs/java/server_msg_pulsar/README.md
@@ -1,0 +1,8 @@
+Athenz Server Messaging/ChangeLog Implementation using Pulsar
+=============================================================
+
+## License
+
+Copyright The Athenz Authors
+
+Licensed under the [Apache License, Version 2.0](http://www.apache.org/licenses/LICENSE-2.0)
